refactor: extract conditions logic to new file
